src/terminator/Terminator.java: implement a small in-application server so that we can respond to "new" requests by opening a new window...
bin/terminator: ...which lets the shell script avoid starting a new Terminator
if we already have one running. There's no need to do this on Mac OS, because
the system already supports this naturally (you start a program from the Dock,
and once it's started, the icon activates the program, and C-N is the usual
way to get a new window). Of course, this will be annoying on other OSes too,
when I'm working on Terminator and want to ensure I'm running a new instance.
I'll fix that when I need to.