post-commit: In response to my previous suggestion, Elliott said (among other less positive things) [
it would also be another nudge in the direction of needing a checked-out
copy of the repository with the scripts, and updating in the script, and
re-running if it detected a change.

]

Well, I've created a checked-out version of the salma-hayek repository in software@jessies.org:~/checked-out/salma-hayek.  (I checked out most of the others too but just deleted them on the grounds that, if I'm not updating them or using them, then leaving them lying around would be misleading.)

This change gets the post-commit trigger to update the checked-out repository that it lives in, after sending the mail.

If it were easy to make it work robustly, I'd have probably implemented the last part of Elliott's suggestion.  So an update to the post-commit trigger gets mailed out by the old post-commit trigger.  The next commit uses the new trigger.  Is that so bad?

So this should be the first mail sent by my modified post-commit.  From the previous mail, I see that my changes to at least one of the symlinks worked OK.