ARGH!  Checkintool threw away my comment for the fifth time, this time on a "Discard changes".
bin/terminator: Fix this problem with the cached_icon file generation:

cat: /home/martind/.terminal-logs/.cached-icon: No such file or directory
/home/martind/software.jessies.org/terminator/bin/terminator: line 65: [: argument expected
cat: /home/martind/.terminal-logs/.cached-icon: No such file or directory

I didn't know that -o could be used in bash as a disjunction - but it can.  In a different section of the bash man page that's talking about the same sort of conditional expressions, I learned that it can also be used to determined whether a named shell option is enabled.  How it decides which meaning to ascribe, I'm not sure.  Anyway, the problem is presumably that the whole "pipline" was expanded before being executed.  Splitting it into a "list" of two "pipelines" seems to fix the problem.

Add the $DISPLAY number to the base port number of 1949.  I don't know where Elliott got that number (he doesn't say in the code or the check-in comment).  A few either side seem free in my /etc/services file.  Pass the number to the Java...

src/terminator/Terminator.java: ... initially because I don't know how to get the $DISPLAY number in Java, especially not in a way that's portable to non-X systems.  Latterly, I realized that this removes a duplicate magic number.  Bon.