TODO: Added some PTextArea TODO stuff.  Unfortunately, by the time I came around to checking in, most of them were done.
src/e/ptextarea/PAnchorRegion.java: Added useful methods called getStartIndex and getEndIndex so we don't have to call getStart().getIndex() etc.

src/e/ptextarea/PTextArea.java: Added the concept of a selection, and a bunch of useful methods whose signatures are copied straight out of JTextComponent (and a few which aren't).  Selection painting is currently done before any other highlights, which seems to be what Edit does.  Moan if this is wrong.  Also added the PMouseHandler as a mouse motion listener.

src/e/ptextarea/PMouseHandler.java: Our mouse handler now understands dragging, and will modify the selection as you do so.

src/e/ptextarea/PKeyHandler.java: Modified the character insertion so it deletes any selected text beforehand.  The delete and backspace keys now delete the selection if there is one, and if not revert to their normal behaviour.  The motion keys now check for Shift being pressed, and modify the selection if it is.