bin/svn-log-to-html.rb: Elliott's worrying me again by talking about automatically putting hard line breaks in check-in comments.  He complains that my check-in comments look too wide to be readable in ChangeLog.html.  The right way to solve that is, I've always presumed, CSS.  So, in the spirit of being constructive, I'll learn a little CSS.
Spurious line breaks can't be removed automatically without detecting quoted code.  That, it seems, requires some kind of mark-up to be reliable, though it could be primitive mark-up.  It would be an pain to insist on mark-up and mistakes would be likely.  You'd have to be able to edit comments when you made mistakes, which is at least possible with Subversion although not (as far as we know) with BitKeeper.

The browser is better at line breaking than a monkey, certainly than a programmer with more important matters on his mind.  It's not unreasonably difficult to implement soft wrapping in a text component used for ChangeLog editing.  So spurious line breaks are unnecessary and detrimental and the alternative is practical.

Anyway, I wasn't much interested in picking a width - I was more interested in learning the mechanism.  This guy's "scalable width demo" convinced me to use "em" in the mechanism:

http://cookiecrook.com/AIR/2003/train/accessiblecss.php

Elliott said "10-13 words per line".  This guy agreed:

http://www.maxdesign.com.au/presentation/em/

So I tried his em suggestions, starting at 30.  Even when I made the width 40em, it mangled Elliott's comment in revision 452 of the terminator log.  45em didn't look like it mangled anything and "looks good to me".

Surfing random pieces of research suggests that lots of the quoted results optimise the speed of reading newsprint by the general public.  That doesn't necessarily generalise to a more specialist application.

This site has a useful summary of what to do about a number of issues, including words per line, font size, font choice and choice of colors:

http://psychology.wichita.edu/optimalweb/text.htm

This little piece of research was actually into web pages rather than newsprint and seems to conclude that the problem is more complicated than just one of words per line - that margin width may significantly vary the optimal number of words per line:

http://hubel.sfasu.edu/research/textmargin.html