TODO: Extra job to be done (unrelated to the rest of this checkin).
src/e/ptextarea/PTextArea.java: Added the idea of a selection anchor, and some methods to manipulate the unanchored end of selection.  This puts all the 'which end do I choose?' logic which was spread around the key and mouse handlers into one place.  Allows us to do the job properly, too.

src/e/ptextarea/PMouseHandler.java: Used the new anchored/unanchored API in PTextArea to simplify the code.  We no longer need to remember which end is the anchor ourselves.  It does slightly complicate the double-click-drag case, since the 'anchor' is actually around a whole word, rather than a fixed position.  But the real gains are found in...

src/e/ptextarea/PKeyHandler.java: ...where we can rip out a whole load of stuff which was trying to work out which end of the selection to change.  Now we know - we just ask the PTextArea for the unanchored end.  If there's no selection, this'll be the caret location, so everyone's happy.