bin/mailpatch.rb: post-commit was already doing a more tasteful job than BlueArc's mailpatch.
post-commit: So extract the bits which aren't svn-specific...

patch-to-html-email.rb: ... into somewhere which mailpatch.rb can share.  Having another application makes it worth spending a little more effort on:

I've got rid of the <pre> tags, used <tt> around the +, - or space line prefix and &nbsp; for the indentation, fixed the HTML compliance to some degree, parsed the file diffs one-at-a-time and added an ASCII art detector to wrap each file in <tt> if the patch contains ASCII art.