src/e/edit/FindAndReplaceAction.java: Using the same font in find/replace dialog as in the associated text sounds eminently reasonable.  The motivation was editing an ASCII art table and, previously, having to guess or count the number of spaces required to line up the columns.  Now, ASCII art files get a fixed width replace preview.
The need for two calls to setFont is perplexing.  If the call in the constructor is omitted, then single line replace lists get the old, default font.  If the call in getListCellRendererComponent is omitted, then large replacement lists get the old, default font.  If just the latter call is included, then even the first item is rendered in the right font.

I'd rather not check-in with the code in such a poorly understood state but I want to see if my patch comes out shiny in the new post-commit trigger.