src/e/ptextarea/PBracketUtilities.java: More fool me, having poked fun at Elliott for being a part-time C++ programmer, not to see the +-1 error on the other side of the colon.  Mind, fixing that didn't actually fix the problem I documented last time.  That was due to the last "offset != stop" guard being too weak.  The test should have been "nesting == 0".
If we're going to use the STL idiom, then let's use it fully.  Everything will be clearer then.  That means sticking rigorously to declaring the loop variable in the for() and breaking or, preferably, returning out of the body of the loop.  If we do the return test in the right place, then we avoid the need to start one ahead of where the idiom tells us.  That replaces a 1 with a 0 and removes the duplicate loop increment.  It also means that we don't continue one step too far, meaning that the iterator doesn't get to the "stop", meaning that the previously erroneous guard would work.

Now we don't have access to the iterator outside the loop, we have to decide what to return.  That could have been a problem except that the comment above the function tells us what we should be returning.  We were never returning -1 before.  I wonder if it will cause any problems?  It didn't in my testing.

While I'm looking at the comment, fix it so it doesn't lie that we're finding the offset of the matching bracket.  We're finding the offset *inside* the matching bracket, which is different when we're searching backwards.  I expect someone is going to trip over this again if we leave the name as "findMatchingBracket" but perhaps all bracket matching code wants the "inside" offset.

Speaking of comments, if one and a half C++ programmers tripped over the setting of "stop", it needs a comment.  In the termination comment, remove the pleonastic "actually" and "just" and replace the ambiguous "before" with "inside".  The ? 1 : 2 asymmetry was, I imagine, what provoked the comment in the first place.  The asymmetry deserves a line and comment of its own.  Now we're doing the test in the right place, we've got rid of half of the conditional.  That's good because that one line was doing two things.