src/e/ptextarea/PNewlineInserter.java: Fix the long standing irritation whereby open-brace, Enter in the middle of a line moves the following text after the closing brace.  So pressing Enter after the brace here:
  hello {world

would lead to:

  hello {
    
  }world

instead of the desired:

  hello {
    world
  }

(Good lord, checkintool's check-in comment area has a matching brace inserter.)

I *think* I've tested this pressing Enter after an open brace which is:

o  the last character in the buffer
o  the last character in the buffer bar a newline
o  the first character in the buffer
o  on the end of the first line in the buffer
o  on a line with no indent
o  on a line with indent
o  in the middle of the last line in the buffer
o  in the middle of the last line bar a newline

No alarms and no surprises, unlike the pre-Evergreen version, which tripped over the getLineEndOffset end-of-buffer hack in JTextArea.  Score one for Phil's BeforeTerminator semantics.