src/e/ptextarea/PBracketUtilities.java: fixes bracket highlighting in the following cases, without breaking double-click bracket matching (each string represents an entire document, and remember to test each bracket separately in those strings that have two): "(hello) ", " (hello)", "hello)", "(hello".
src/e/ptextarea/PMatchingBracketHighlighter.java: now findMatchingBracket is
returning what the comment always claimed, this caller can be simplified. Also
highlight brackets that failed to match in a different color, rather than
making you hunt for a non-existent match. [If you're wondering how one caller
can change but not the other, this patch actually breaks double-click bracket
matching when moving backwards, because -- unlike most editors -- we select
inside the brackets, and don't include the brackets themselves. I'm not sure
what the most tasteful fix to the mouse handler is, so I'll leave that for
a later patch because I am sure that findMatchingBracket should always give
you the offset of a bracket.]