src/e/ptextarea/PMouseHandler.java: Fix double-click bracket matching in the direction where you click on a closing bracket, to select inside the bracket.  While elegant, in that the code is concise, the comment doesn't really illuminate matters so much as make the reader stop and work it out for themselves.
Elliott said "ah, i see: the +1 is ignored in the forwards case" [and is the active ingredient in the desired, superior, behaviour in the backwards case].

Elliott suggested that we might want to improve the bracket-matching API on the input parameter side too, by insisting on being given the offset of a bracket.  At the moment, we're given offset+1 of an opening bracket.  That just makes the two callers symmetrical in terms of having to pay more attention to what they're doing.  It would make PBracketUtilities less specific to PTextArea's unconventional, but superior, behaviour.