src/terminator/Terminator.java: Avoid calling System.exit.  The JavaDoc says it's a bad idea and, indeed, it proves to be so when the InAppServer asks us for our --help or --version and causes all the Terminator windows to disappear as if the program had crashed.  (Fortunately, I wasn't doing anything in the one Terminator window I had open, so I was more amused than cross, but this needed fixing before it bit anyone.)  Instead:
Take care not to initialise the UI or start any background threads until we know that we're not just being asked for --help or --version.  Background threads stop the JVM from quitting if we fall out of main.  Write the --help or --version into the InAppServer's socket rather than, uselessly, into the original process's stdio.  (With my setup, that ends up in ~/.xsession-errors-`date --iso`.)  I don't much like the parseOriginalCommandLine hack.

We can write the --help or --version into the InAppServer's socket.  My first attempt at fixing this problem was to modify the Ruby script not to use the InAppServer if it saw options on the command line.  That might have been useful but I failed to bend Ruby to my will.  In any case, the Java benefits from being more elegant, in having fewer references to System (even if my ham-fisted pidgin efforts make it less elegant elsewhere).  Being able to interrogate the options and version of the currently running terminator might turn out to be useful.

src/terminator/Options.java: InAppServer gives us a PrintWriter.  The JavaDoc suggested that a PrintWriter is the alternative to choose if we're writing characters rather than bytes.

bin/terminator: This stops the --help and --version output coming out double spaced, when it comes via the InAppServer.  Ruby readlines() gives you the terminator on each line.  Not how you'd have done it, not how Phil'd have done it but there you are.  There's a copy of this code in Edit that I'll check-in in a minute.