src/e/ptextarea/PCppIndenter.java: Pre-Evergreen Edit said:
                fileType = C_PLUS_PLUS;
                text.setIndenter(new JavaIndenter());

From a quick inspection, I think there's nothing in salma-hayek's JavaIndenter which doesn't also apply to C++ (and, probably, C99).  This fixes the vague "the indenter doesn't work" complaint that I verballed the other day and, specifically, the one instance which Elliott happened to observe over my shoulder.

The default indenter doesn't realize that braces signify a new level of indent.  Edit *does* think that an Enter after a mismatched brace should cause extra indentation.  This is hard-wired in PNewlineInserter, although within a guard that the indenter considers closing braces to be electric.  The default indenter considers closing braces to be electric.