src/e/ptextarea/PJavaIndenter.java: Separate the parts which aren't wanted in all the derived classes, in preparation for a rearrangement.
I noticed that getActivePartOfLine wouldn't do the right thing with trailing spaces if they're before a C++-style comment.  Reorder the comment stripping and trimming to fix that.  Separate the C++-style comment stripping as it only applies to Java and C++ (and C99), not Perl.

Guard the label recognition as I don't think it applies to Perl.  Not to my deliberately pidgin Perl, anyhow.  I know that bash has an odd case syntax and I'm not sure I want to find out what Perl's is.  YAGNI.

Guard the doc-comment continuation code as I know that doesn't apply to Perl.  It's not such a strong convention in C++ as it is in Java but it's just as nicely readable there.

Only "case" is followed by anything other than spaces and a colon.  Fix the regular expression while splitting them to remove the C++-specifics from Java.