src/e/ptextarea/PCppIndenter.java: src/e/ptextarea/PPerlIndenter.java: The C++ and Perl indenters aren't indenting Java, so extend a class with a more appropriate name.
src/e/ptextarea/PJavaIndenter.java: The Java indenter preserves the previous behaviour.  This is clear from the diff because I used svn cp.

src/e/ptextarea/PSimpleIndenter.java: Provide a default implementation of all the behaviour which can be overridden.  This is one option.  The better - making all the methods abstract - I'll have to look-up the syntax of.  I wanted to get something consistent checked-in quickly.