src/e/ptextarea/PSimpleIndenter.java: src/e/ptextarea/PNewlineInserter.java: Treat other brackets like we treat braces.  This Perl was the motivating example:
my @options = ( 
  "delete", 
  "links", 
  "recursive", 
  "times" 
);

It's also the first step (in the direction I've chosen) towards better support for inserting a newline at the end of this Java idiom (the pipe character marks the insertion point):

new Thread(new Runnable() {|

Where you'd like to end up with:

new Thread(new Runnable() {
    |
});

On reflection, there's a lot of C code too which opens a long list of parameters with a line ending in open parenthesis and then declares one parameter per line.  We've never liked that style much (regarding long lines as a mistake) but it now seems contrary to deliberately prevent better support for that by only indenting on braces.  If we never hit Enter after an open bracket, then it's not like we're gaining anything by not doing auto-insertion on brackets.  Elliott may doubt that but I'm confident that it'll be fine and think he should try it.

src/e/ptextarea/PIndenter.java: Explain that isElectric doesn't return such a definite result (for Ruby).  Explain why that still makes it useful, albeit not for all the uses it was being put to.