src/e/ptextarea/PNewlineInserter.java: When counting whether there's a bracket mismatch, it's not much more work to keep track of which brackets are mismatched, how many and in which order.  Well, OK, it's a fair bit of work but it wasn't too hard to get something that seems to work plausibly well.
I've had to special-case the '<' and '>' brackets.  At the moment, we remove comments and string literals with regular expressions.  We should use the Styler or the segment iterator or whatever it is as our one true parser.  Then it might be worth having an attempt at distinguishing template<> brackets from the less than token (and <= and << and <<< and <<=).  I'm not sure the C++ case will be practically solvable.

This means that the newline inserter doesn't insert a matching '>' if you insert a newline after '<'.  The indenter will still increase the indent after seeing a line ending in '<'.  A manually inserted '>' at the start of a line will electrically outdent.

I think that's going to be fine.  If a newline is inserted at the pipe:

new Thread(new Runnable() {|

This code will leave you with:

new Thread(new Runnable() {
    |
})

This just needs some work on the code which decides on the need for a semi-colon to get us to a point where I think Elliott will be happy, even if he doesn't think so.  I'm wondering how to detect the difference between:

int functionWithTooManyArguments (|

... which might or might not require a semicolon, depending on whether it's a declaration or definition, and:

new Thread(new Runnable() {|

... which definitely does require a semicolon.