TODO: Job done.
src/e/ptextarea/PNewlineInserter.java: Could do with tidying up and testing but that dropped out pretty nicely.  This fixes the motivating case, as stated in the TODO and:

int functionWithTooManyArguments(|P0 a0, P1 a1, P2 a2);

Which becomes:

int functionWithTooManyArguments(
    |P0 a0, P1 a1, P2 a2
);

If you miss the semicolon, as Elliott did in one email (but not in the TODO), then:

    new Thread(new Runnable() {|)

Becomes:

    new Thread(new Runnable() {
        |)
    });

Which, I submit, is pretty benign, leaving you next to the character which "obviously" wants correcting.