Changes related to fixing this bug [
Check a file in checkintool then its name, which appears in the check-in comment area, is now surrounded by a box whose purpose, if any, I don't understand.  If I click, the box disappears, the caret appears and everything works fine.  If I just start typing: 

2005-04-28T20:07:09.843-0700 CheckInTool: Exception occurred during event dispatching. 
Associated exception: 
java.lang.NegativeArraySizeException 
            at e.ptextarea.PTextBuffer.copyChars(PTextBuffer.java:194) 
            at e.ptextarea.PTextBuffer.replace(PTextBuffer.java:255) 
            at e.ptextarea.PTextArea.replaceRange(PTextArea.java:282) 
            at e.ptextarea.PTextArea.replaceSelection(PTextArea.java:287) 
            at e.ptextarea.PKeyHandler.insertCharacter(PKeyHandler.java:149) 
            at e.ptextarea.PKeyHandler.keyTyped(PKeyHandler.java:53) 
            at java.awt.Component.processKeyEvent(Component.java:5443) 
            at javax.swing.JComponent.processKeyEvent(JComponent.java:2680) 
            at java.awt.Component.processEvent(Component.java:5265) 
            at java.awt.Container.processEvent(Container.java:1966) 
            at java.awt.Component.dispatchEventImpl(Component.java:3955) 
            at java.awt.Container.dispatchEventImpl(Container.java:2024) 
            at java.awt.Component.dispatchEvent(Component.java:3803) 
            at java.awt.KeyboardFocusManager.redispatchEvent(KeyboardFocusManager.java:1810) 
            at java.awt.DefaultKeyboardFocusManager.dispatchKeyEvent(DefaultKeyboardFocusManager.java:668) 
            at java.awt.DefaultKeyboardFocusManager.preDispatchKeyEvent(DefaultKeyboardFocusManager.java:916) 
            at java.awt.DefaultKeyboardFocusManager.typeAheadAssertions(DefaultKeyboardFocusManager.java:794) 
            at java.awt.DefaultKeyboardFocusManager.dispatchEvent(DefaultKeyboardFocusManager.java:632) 
            at java.awt.Component.dispatchEventImpl(Component.java:3841) 
            at java.awt.Container.dispatchEventImpl(Container.java:2024) 
            at java.awt.Window.dispatchEventImpl(Window.java:1766) 
            at java.awt.Component.dispatchEvent(Component.java:3803) 
            at java.awt.EventQueue.dispatchEvent(EventQueue.java:463) 
            at java.awt.EventDispatchThread.pumpOneEventForHierarchy(EventDispatchThread.java:234) 
            at java.awt.EventDispatchThread.pumpEventsForHierarchy(EventDispatchThread.java:163) 
            at java.awt.EventDispatchThread.pumpEvents(EventDispatchThread.java:157) 
            at java.awt.EventDispatchThread.pumpEvents(EventDispatchThread.java:149) 
            at java.awt.EventDispatchThread.run(EventDispatchThread.java:110) 

] about which Elliott said [

my (unconfirmed) suspicion is that there's something wrong with the anchor stuff; in this case, we can end up with a PAnchorRegion where end < start (because that's only checked on initialization, and the two can be mutated independently). it might also explain why the spelling checker's highlights wander about as you edit.

on the other hand, this CheckInTool problem only appeared with my patch to make append not modify the selection. i assume that i just brought a latent problem to light. 

].  I think he was correct but I only understood what he was talking about after working out for myself what all the relevant classes do.

src/e/ptextarea/PAnchorRegion.java: I thought that someone might be mutating the anchor after construction via the getStart or getEnd methods here.  It turned out that no-one needed to use them...

src/e/ptextarea/PTextAreaSpellingChecker.java:
src/e/ptextarea/PTextArea.java: 
src/e/ptextarea/PHighlight.java: ... because all the callers are only interested in getStartOffset and getEndOffset.

src/e/ptextarea/PAnchorSet.java: The problem turned out to be the "return mid" if we found a matching index.  We need to return the first anchorSet index which matches the textIndex, rather than the first one we find.  textInserted and textRemoved depend on this to update all of the anchors after (and including) the one at the index returned.  As well as fixing the problem, I've renamed the function to make it less likely that we repeat the error.  I've renamed "min" and "max" as "start" and "end" to indicate that we're using them as a half-open region (ie inclusive start, exclusive end).

I made the fix in the first way which occurred to me.  I did think about it some more and, although it was clear that there must be a more elegant way, it wasn't clear in which direction that lay.  It's distressing that we're reimplementing binary chop in 2005.  Elliott says that we probably only need to do this because of the WeakReferences and that we can probably remove the WeakReferences and explicitly delete the PAnchors.  It looks like PAnchor is only used by RegionAnchor by PAnchorRegion by PHighlight and that we already need an explicit delete method on PHighlight.  So I think he's right but we could do with some agreement or disagreement from Phil.