src/e/edit/AppropriateFontAction.java: There was no obvious way to ask Edit to reassess the file's contents for ASCII art.  I can think of two non-obvious ways:
o  select and copy the class name into the clipboard
   close the file
   Open Quickly
   paste the class name
   resolve any ambiguity
   reopen the file
   find your place again

o  save the file
   make a spurious change
   tell Edit to Revert to Saved
   confirm that you want to discard the changes

I was tempted to say that every time we update the watermark, we could also reassess the font but that would just be a hack.  Maybe we should just get rid of all such hacks and update the watermark, font etc on focus changes, external tool actions that the user has asked for and every so often on the focussed file.  On the other hand, external tool actions only modify the on-disk state and the appropriate font is determined solely by the in-memory version.

src/e/edit/EditMenuBar.java: Add the new item to the menu and cope with...

src/e/edit/FixedFontAction.java: 
src/e/edit/ProportionalFontAction.java: 
src/e/edit/ChangeFontAction.java: ... the refactoring of the existing font actions to a more extensible style.