native/all/libpty/Makefile: Remove all references to the specific library being built here.  Use a more declarative style which tends to scale better than ifeq.  Use := for external tool invocation so the tool's only invoked once.  Use make's built-in compilation rules, modifying the normal variables like CPPFLAGS instead of taking over the whole rule.  Build object files in a target-dependent directory, which makes "make clean" easy and thorough.
Tested on (duezer|ted).us.dev.bluearc.com (Linux and Darwin).

This file's missing the boilerplate from java.make but, other than that, it should be ready to be moved to salma-hayek and renamed jni.make.  I don't think we should do that.  I think we should have a universal.make instead.  Which is to say that this would be merged with java.make.