native.make: Add a rule for building the JNI header.  Work out whether we're building JNI by whether we have a source file whose name includes an underscore.  Make will work out that it needs to link the source into the $(GENERATED_DIRECTORY).  We just need to tell it to link the headers.  Be more intention-revealing about that.
The Java is cleaned and recompiled from scratch on every invocation of top-level make.  This will cause the JNI header to be rebuilt which will cause the .cpp file to recompile.  We've fixed such sub-optimal situations before and can do so again if it becomes a problem.  It won't, so I've no immediate intention of doing anything about it.

This breaks the Mac OS X build unless JAVA_HOME is defined.  Elliott said that'd be OK.  I'm not sure whether I can work out how to default JAVA_HOME based on what's already in java.make.  Perhaps.