Share the common code between java.make and native.make as variables.make and rules.make, rather than copying it.  This is effectively my admission that universal.make is still a way off.  The interesting part will be supporting multiple native targets from one makefile.  I still think that's worthwhile but it'll be difficult to come up with changes to make that work without a motivating example with which to test it.  At the moment, we only have one native target per repository.
native.make: If this file is going to live a while, we'll want the comment which marks the dividing line between variables and rules.  Violating that dictum causes complexity and obscure bugs.

java.make: The immediate motivation for this ChangeSet was to be able to use the $(SOURCE_EXTENSIONS) from native.make to locate unchecked-in source under native.  Elliott raised the valid concern that, if we were searching for subdirectories to compile via their .svn directories, then we'd have to check-in anything before we could compile it for the first time.  That'd be a helpful constraint in a distributed version control system like BitKeeper.  It's often frustrating at work to find that a peculiar line sprang into life fully form in the first revision.  In Subversion, however, it'd break everyone else's build.

variables.make: The default compilation rules are suffix rules so, while disabling .SUFFIXES: is appropriate for BlueArc, it's not appropriate where we want to use the default compilation rules - in native.make.

rules.make: We could share more rules.  This is the only one which was previously shared.