java.make: Remove the trailing slash on the $(SUBDIRS) names which $(dir) deliberately leaves behind.  This seemed to be causing NSSpell not to build on Mac OS X because $(basename) of a name ending in slash is the empty string there.
How did I not notice this before my previous check-in?  I suppose I'm willing to ascribe that to my mistake, although I've scrolled up the terminator window and I can see the make clobber; make.  More interestingly, though, why doesn't it cause me a problem on Linux?  Am I to take it that all make-3.80s are not the same?