.: Ignore the top-level generated directory.
Does the panel think that my hacked Subversion.java will pass --non-recursive to svn commit and that svn commit will do what I want?  Rather than what it says on the tin.  What I *want* is for it not to commit my native.make change because I haven't got the terminator script change ready.  What it says on the tin is that it just won't recurse into subdirectories.  When I do "svn diff --non-recursive .", it does what it says on the tin.  When I do "svn commit --non-recursive .", it seems like it's going to do what I want (according to the comment it generates in vim).