native.make: Just to keep the whining down, generate the JNI header in a different directory.  Only copy it to the compilation directory if there isn't an identical one there already.  I'd been wondering when I've have to surround the JNI rule with a conditional.  I expected that it would be the first time when two non-JNI native directories were compiled and the rules clashed.  The "build:" dependency has fortunately provoked the issue earlier.
While investigating that Mac OS X make crash earlier, I looked at the make NEWS file for the first time.  Under 3.80, it mentioned a feature that I hadn't heard of before: "order-only prerequisites".  In some BlueArc makefiles, we say "generated-files: trigger-file;" which says that updating the trigger-file makes the generated-files out-of-date.  This isn't at all what we mean - we just want it to generate the trigger-file first then notice that the generated-files have magically sprung into existence.  It does seem to work without causing lots of spurious recompilation but I've always had my doubts as to whether it'll keep working.  An "order-only prerequisite" is exactly what we want.

That applies here too.  We don't want the $(COMPILED_JNI_HEADER) to be deemed out-of-date when the $(GENERATED_JNI_HEADER) is regenerated but we do want the $(GENERATED_JNI_HEADER) to be regenerated first, in case the rule updates $(COMPILED_JNI_HEADER) as a side-effect.

The compilation didn't take very long anyway and this complicated rule is now generating more output than it was before.  So I fully expect more whining.  I'll happily add an "@" to the front of the rule when that happens.