java.make: native.make: Own up that we only support building JNI libraries (and executables) at the moment, not general shared libraries.
native.make: There's no need to decide on a default target, as that's java.make's problem.  Instead of insisting that JNI directories start with "lib", we may as well use the same condition as we use elsewhere: whether we find any plausible-looking JNI source.  (Which, at the moment, means source with an underscore in the filename.)