java.make: native.make: Support building on cygwin.  That involves the following differences:
Everywhere we pass filenames to Sun tools, we have to convert them from cygwin format to Win32 format.

The Win32 classpath, like the Win32 PATH, is semicolon separated.  That provides additional challenges when $(shell) goes via bash.

The normal place for $(JAVA_HOME) seems to be "C:/Program Files/Java/jdk1.5.0_03".  Makefiles traditionally abhor filenames with spaces.  We could insist on using Progra~1 (that works) but I'll make an effort to cope with spaces for the minute.

The shared library needs to be "pty.dll".  The "libpty" directory should be renamed "pty".  The "lib" prefix is an artifact of a Unix convention, not a Java one.