java.make: Fix running terminator after "make clean".  Previously that said things like:
> make: *** No rule to make target
> `/private/automount/home/elliotth/Projects/terminator/generated/
> native/all/pty/Darwin/libpty.jnilib'.  Stop.

When it really meant "no rule to make the .class file, to make the JNI header, to make the JNI library".  (make has no choice but to be so unspecific in its error.  The user's intention might have been to generate the target some completely different way, for example from SCCS/s.libpty.jnilib.)