src/e/ptextarea/PMouseHandler.java: Revert the previous change (including the spurious whitespace at the start of the file), so that the middle button again pastes my Windows clipboard, as transferred only to the X11 selection, by the wonder of x2vnc...
src/e/ptextarea/PTextArea.java: ... while leaving Ctrl-V and the paste menu item solely bound to the system clipboard, so that anyone who never uses the X11 selection should never notice it's there.

It will certainly be a pleasant change to have Ctrl-V and the middle button pasting operating in a simple, consistent manner - with Ctrl-V just operating on the clipboard and the middle button just pasting the selection.  No one yet countered my suggestion that terminator should do the same.

I'm more unimpressed with Elliott's and my inability to write three trivial lines of correct code than by the design mistake that is the X11 selection.  It looks like xvncviewer, which claims only to be 3.3.7-7, sets the X clipboard as well as the selection.  So anyone setting out to derive a new version of x2vnc from that would fix this problem the right way (for that particular application).  Perhaps the clipboard handling code could be lifted and retro-fitted to the existing x2vnc.  The maintainer is quite amenable to patches.  That would require a lot more effort than this patch.