src/e/edit/FindFilesContainingSelectionAction.java: the selected text is never null with PTextArea, it's the empty string instead.
src/e/edit/Workspace.java: we ought to re-use the find files dialog like we
reuse all the others. These changes fix the find files dialog so that it
remembers what you last searched for. We probably broke it when we made
ETextArea override JTextArea's getSelectedText to not return null, but didn't
notice the problem soon enough to connect the two.