src/e/edit/FindAndReplaceAction.java: The "More than 100 matches" problem has been demonstrating itself reproducibly since my EMonitoredTextField change.  I've been looking at SwingWorker.  I was (and am) mightily confused by how SwingWorker can have a "get" and a "getValue".  "finished" here was the only caller of "get".  get's function seems to be to wait for the worker thread to have quit and to return its result.  The only caller of "finished" is the invokeLater which only happens when the worker has completed.  Completed but not necessarily when it has quit.  The invokeLater callback could be executed while the worker thread is still terminating.  I don't see the value in holding up the finished callback that little bit longer as nothing is waiting on the finished callback.
That's, incidentally, what causes the "100 matches" problem.  Because a second worker is started before the finished callback has completed, we end up with two finished callbacks being on Swing's queue at once.  It seems to execute them in LIFO order, at least sometimes (so probably always).

Anyway, I don't see the reason for using "get" instead of "getValue", like findFilesDialog.

I may be misunderstanding Java but I don't see how the while loop in showMatches can ever be executed more than once.  Perhaps the intention was to spin-wait for the worker thread to complete.  That doesn't seem like a good idea for the CPU usage.  Perhaps the intention was to wait for the finished callback to complete.  That would just lead to deadlock - the finished callback has to be executed on the same thread (the Swing thread) as the one which calls showMatches.  Without a comment, this loop looks like a mistake, so out it comes.

SwingWorker appears to need some work.  I don't see any reason why the finished method shouldn't be abstract.  A method as complicated as "get" shouldn't have such a commonplace name.  get's usage of Thread.currentThread().interrupt() can't possibly do anything useful.